refactor: move loadCompletions from snippet to completion module
- Extract loadCompletions function to src/completion/settings.ts
- Fix module boundary violation (completion logic in snippet module)
- Add completion source cache with functional implementation
- Update imports in completion/source/index.ts
- This creates clearer module boundaries and responsibilities
