+++ src/completion/source/index.ts
@@ -1,19 +1,26 @@
-import { loadCompletions } from "../../snippet/settings.ts";
+import { loadCompletions } from "../settings.ts";
 import type { CompletionSource } from "../../type/fzf.ts";
 import { gitSources } from "./git.ts";
-import { ZENO_DISABLE_BUILTIN_COMPLETION } from "../../settings.ts";
-
-let cachedCompletionSources: readonly CompletionSource[] | undefined;
+import { getEnv } from "../../config/env.ts";
+import { getCompletionSourceCache } from "./cache.ts";
 
 export const getCompletionSources = async (): Promise<
   readonly CompletionSource[]
 > => {
-  if (!cachedCompletionSources) {
-    const userCompletions = await loadCompletions();
-    cachedCompletionSources = [
-      ...userCompletions,
-      ...(ZENO_DISABLE_BUILTIN_COMPLETION ? [] : gitSources),
-    ];
+  const cache = getCompletionSourceCache();
+  const cached = cache.get();
+
+  if (cached) {
+    return cached;
   }
-  return cachedCompletionSources;
+
+  const env = getEnv();
+  const userCompletions = await loadCompletions();
+  const sources = [
+    ...userCompletions,
+    ...(env.DISABLE_BUILTIN_COMPLETION ? [] : gitSources),
+  ];
+
+  cache.set(sources);
+  return sources;
 };
